pyproject.toml
[tool.poetry]
name = "ieee-rero"
version = "0.1.0"
description = ""
authors = ["RohanJnr <rohanjnr44@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
python-dotenv = "^1.0.0"
google-cloud-firestore = "^2.12.0"
firebase-admin = "^6.2.0"
pylint = "^3.0.0"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
